==> germcells/2_CCAalignClusteringSPG.Rmd <==
# CCA alignment and clustering for SPGs from three species

##read in 1-1-1 orthologues data and spg raw data.
```{r}
humMacMouOrth.1to1.in <- readRDS("../interdata/humMacMouOrth.1to1.in.Rds")
dge.hum.spg.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.hum.spg.raw.rds")
dge.mac.spg.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mac.spg.raw.rds")
dge.mou.spg.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mou.spg.raw.rds")
library(Seurat)
library(seriation)
library(xlsx)
```

##Seurat CCA
###create mouse seurat object
```{r}
dge.mou.spg.raw.sorted.named <- dge.mou.spg.raw[humMacMouOrth.1to1.in$mmusculus_homolog_associated_gene_name,]
rownames(dge.mou.spg.raw.sorted.named) = humMacMouOrth.1to1.in$external_gene_name.x
dge.mou.spg <- CreateSeuratObject(raw.data = dge.mou.spg.raw.sorted.named)
dge.mou.spg <- NormalizeData(object = dge.mou.spg, normalization.method = "LogNormalize", 
    scale.factor = 10000)
dge.mou.spg <- ScaleData(object = dge.mou.spg,vars.to.regress = c("nUMI"))
dge.mou.spg <- FindVariableGenes(object = dge.mou.spg,mean.function = ExpMean, dispersion.function = LogVMR, x.low.cutoff = 0.0125, y.cutoff = 0)
```

###create monkey seurat object
```{r}
dge.mac.spg.raw.renamed <- dge.mac.spg.raw
rownames(dge.mac.spg.raw.renamed) <- rownames(dge.hum.spg.raw)
dge.mac.spg.renamed <- CreateSeuratObject(raw.data = dge.mac.spg.raw.renamed)
dge.mac.spg.renamed <- NormalizeData(object = dge.mac.spg.renamed, normalization.method = "LogNormalize", scale.factor = 10000)
dge.mac.spg.renamed <- FindVariableGenes(object = dge.mac.spg.renamed,mean.function = ExpMean, dispersion.function = LogVMR, x.low.cutoff = 0.0125, y.cutoff = 0)
dge.mac.spg.renamed <- ScaleData(object = dge.mac.spg.renamed,vars.to.regress = c("nUMI"))
```

###create human seurat object
```{r}
dge.hum.spg <- CreateSeuratObject(raw.data = dge.hum.spg.raw)
dge.hum.spg <- NormalizeData(object = dge.hum.spg, normalization.method = "LogNormalize", 
    scale.factor = 10000)
dge.hum.spg <- FindVariableGenes(object = dge.hum.spg,mean.function = ExpMean, dispersion.function = LogVMR, x.low.cutoff = 0.0125, y.cutoff = 0)
dge.hum.spg <- ScaleData(object = dge.hum.spg,vars.to.regress = c("nUMI"))
```

##Select Highly variable genes from three species and choose the union.
```{r}
hvg.spg.1 <- rownames(x = head(x = dge.hum.spg@hvg.info, n = 2000))
hvg.spg.2 <- rownames(x = head(x = dge.mac.spg.renamed@hvg.info, n = 2000))
hvg.spg.3 <- rownames(x = head(x = dge.mou.spg@hvg.info, n = 2000))
hvg.union <- union(x = hvg.spg.1, y = hvg.spg.2)
hvg.union <- union(x = hvg.union, y = hvg.spg.3)
```

```{r}
dge.hum.spg@meta.data[,"species"] <- "Human"
dge.mac.spg.renamed@meta.data[,"species"] <- "Monkey"
dge.mou.spg@meta.data[,"species"] <- "Mouse"
```

###Run seurat CCA 
```{r}
spg.cca <- RunMultiCCA(object.list = list(dge.hum.spg, dge.mac.spg.renamed, dge.mou.spg),genes.use = hvg.union, num.ccs = 15, niter = 20)
```

```{r}
MetageneBicorPlot(spg.cca, grouping.var = "species", dims.eval = 1:15, display.progress = TRUE)
```

##Align CCs and run dimensional reduction methods
```{r}
#spg.cca <- CalcVarExpRatio(object = spg.cca, reduction.type = "pca", grouping.var = "species", dims.use = 1:8)
spg.cca <- AlignSubspace(object = spg.cca, reduction.type = "cca", grouping.var = "species", dims.align = 1:8)

spg.cca <- RunPCA(object = spg.cca, reduction.use = "cca.aligned", dims.use = 1:8)
spg.cca <- RunTSNE(object = spg.cca, reduction.use = "cca.aligned", dims.use = 1:8)
spg.cca <- RunUMAP(object = spg.cca, reduction.use = "cca.aligned", dims.use = 1:8)
```

### plot post-aligned cells using t-SNE, UMAP and CCA.
```{r fig.width=5, fig.height=4}
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.species.tsne.pdf",width = 5,height = 4)
TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "species",cells.use = spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dev.off()
DimPlot(object = spg.cca, reduction.use = "cca.aligned", pt.size = 1, group.by = "species")
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.species.umap.pdf",width = 5,height = 4)
DimPlot(object = spg.cca, reduction.use = "umap", pt.size = 1, group.by = "species",cells.use = spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dev.off()
```

##Do clustering
```{r}
spg.cca <- FindClusters(object = spg.cca, reduction.type = "cca.aligned", dims.use = 1:8, save.SNN = TRUE, resolution = 0.6)
```

```{r fig.width=5, fig.height=4}
TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "res.0.6")
DimPlot(object = spg.cca, reduction.use = "umap", pt.size = 1, group.by = "res.0.6")
DimPlot(object = spg.cca, reduction.use = "cca.aligned", pt.size = 1, group.by = "res.0.6")
```

change the label of the clusters
```{r}
spg.cca@meta.data$res.0.6.ord <- plyr::mapvalues(spg.cca@meta.data$res.0.6, c(4,2,0,1,5,3,7,6),c(1,2,3,4,5,6,7,8))
```


Cluster centroid for SPG
```{r}
spg.cen.0.6 <- c()
clus <- unique(spg.cca@meta.data$res.0.6.ord)
for(i in clus){
  cen <- apply(spg.cca@data[,spg.cca@meta.data$res.0.6.ord == i], 1, mean)
  spg.cen.0.6 <- cbind(spg.cen.0.6, cen)
}
colnames(spg.cen.0.6) <- clus
```

```{r}
library(seriation)
```

## cluster rank correlation (clusters 1-6)
```{r fig.width=6,fig.height=5}
#library(ComplexHeatmap)
spg.cor.0.6 <- cor(spg.cen.0.6[,c("1","2","3","4","5","6")], method="spearman")
a = seriate(spg.cor.0.6,method="BEA_TSP")
ord = get_order(a)
Heatmap(spg.cor.0.6[ord,ord], cluster_rows = FALSE, cluster_columns = FALSE)
```

```{r}
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cor.heatmap.pdf",width = 5,height = 4)
Heatmap(spg.cor.0.6[ord,ord], cluster_rows = FALSE, cluster_columns = FALSE)
dev.off()
```

```{r fig.width=5, fig.height=4}
#pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.species.tsne.pdf",width = 5,height = 4)
#jpeg("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.species.tsne.jpeg",width = 5,height = 4, units = "in",res = 600)

pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.species.tsne.pdf",width = 5,height = 4)
TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "species",cells.use = spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)], colors.use = c("#e3a65a","#5a97e3","#e35a97"))
dev.off()
```


###plot the clusters in TSNE and UMAP
```{r fig.width=5, fig.height=4}
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.tsne.pdf",width = 5,height = 4)
TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "res.0.6.ord", colors.use = cols, cells.use = spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dev.off()
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/plots/spg.cca.umap.pdf",width = 5,height = 4)
DimPlot(object = spg.cca, reduction.use = "umap", pt.size = 1, group.by = "res.0.6.ord", cols.use = cols, cells.use = spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dev.off()
DimPlot(object = spg.cca, reduction.use = "cca.aligned", pt.size = 1, group.by = "res.0.6.ord", cols.use = cols)
```
Save spg.cca object
```{r}
saveRDS(spg.cca, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/spg.cca.rds")
```



##Compare with the mouse SPG clusters in previous paper.
##Read in mouse germ cell 12 clusters
```{r}
germ12clu <- read.table("~/storage/HumanMacaqueMouseCompAnaysis/mouse/MouseAdultST24_12GermClusters_ClusterID.txt")
mouse4clus <- read.table("~/storage/HumanMacaqueMouseCompAnaysis/ClusterID_SPGcellsUMI1k_4clusters.txt",header = FALSE)
```

```{r}
spg.cca@meta.data$mouseclus <- 0
spg.cca@meta.data$mouseclus[spg.cca@cell.names %in% mouse4clus$V1[mouse4clus$V2==1]] = "1a"
spg.cca@meta.data$mouseclus[spg.cca@cell.names %in% mouse4clus$V1[mouse4clus$V2==2]] = "1b"
spg.cca@meta.data$mouseclus[spg.cca@cell.names %in% mouse4clus$V1[mouse4clus$V2==3]] = "1c"
spg.cca@meta.data$mouseclus[spg.cca@cell.names %in% mouse4clus$V1[mouse4clus$V2==4]] = "1d"
#spg.cca@meta.data$mouseclus[spg.cca@cell.names %in% rownames(germ12clu)[germ12clu$x == 2]] = 2
#spg.cca@meta.data$mouseclus[spg.cca@cell.names %in% rownames(germ12clu)[germ12clu$x == 3]] = 3
```

###plot the previous clusters on the new tsne.
```{r fig.width=6, fig.height=5}
#pdf("~/storage/HumanMacaqueMouseCompAnaysis/SPG/MouseSPGCluHightlight.pdf",width = 6, height = 5)
DimPlot(object = spg.cca, reduction.use = "tsne", pt.size = 1, group.by = "mouseclus", cols.use = c("grey","#49adad","#cb5b42","#67a64e","#8d70c9","#b69340","#c8588c"),cells.use = rownames(spg.cca@meta.data)[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
#dev.off()
```
```{r}
table( spg.cca@meta.data$mouseclus,spg.cca@meta.data$res.0.6.ord)
```


```{r}
spg.cca <- SetIdent(spg.cca, ident.use = spg.cca@meta.data$res.0.6.ord)
```

##. Find Markers
###. Conserved markers
```{r}
markers.9clus <- list()
for(i in c(1:8)){
  markers.9clus[[i]] <- FindConservedMarkers(spg.cca, ident.1 = i, grouping.var = "species",min.cells.group=2)
}
```

Write to excel
```{r}
for(i in c(1:8)){
if(i==1){
  write.xlsx2(x = markers.9clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_8Clusters_Markers.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = markers.9clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_8Clusters_Markers.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```

###findmarkers
```{r}
markers.sep.9clus <- list()
for(i in c(1:8)){
  markers.sep.9clus[[i]] <- FindMarkers(spg.cca, ident.1 = i, grouping.var = "species")
}
```

```{r}
for(i in c(1:8)){
if(i==1){
  write.xlsx2(x = markers.sep.9clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_8Clusters_Markers_all.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = markers.sep.9clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_8Clusters_Markers_all.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```

##. SPG marker Plots
```{r fig.width=10,fig.height=10}
spg.markers <- c("PIWIL4", "TSPAN33", "MORC1","ID4","FGFR3","DMRT1","DMRTB1","SOLH2H","SYCP3","SYCP1","SPO11","MEIOB","ACRV1","CATSPER1","CATSPER3","PRM1","PRM2","TNP1","TNP2","ZCWPW1","STRA8","GFRA1")
spg.markers <- intersect(spg.markers, rownames(spg.cca@raw.data))
FeaturePlot(object = spg.cca,features.plot = spg.markers,reduction.use = "tsne",pt.size = 0.5,nCol = 4)
VlnPlot(object = spg.cca, features.plot = spg.markers,nCol = 4,group.by = "res.0.6.ord",point.size.use=0)
```
##. SPG marker RA receptor 
```{r fig.width=10,fig.height=2.5}
spg.markers <- c("RARA","RARB","RARG")
spg.markers <- intersect(spg.markers, rownames(spg.cca@raw.data))
FeaturePlot(object = spg.cca,features.plot = spg.markers,reduction.use = "tsne",pt.size = 0.5,nCol = 4)
VlnPlot(object = spg.cca, features.plot = spg.markers,nCol = 3,group.by = "res.0.6.ord",point.size.use=0)
```
##. Cluster Centroids

###. calculate SPG cluster centroid for each species
```{r}
##Cluster 9 removed because of doublets.
spg.exp.all.data <- exp(spg.cca@data) - 1
hum.spg.cen <- matrix(0,nrow=11023, ncol=8)
mac.spg.cen <- matrix(0,nrow=11023, ncol=8)
mou.spg.cen <- matrix(0,nrow=11023, ncol=8)
for(i in c(1:8)){
  hum.cells <- spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord == i & spg.cca@meta.data$species == "Human"]
  mac.cells <- spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord == i & spg.cca@meta.data$species == "Monkey"]
  mou.cells <- spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord == i & spg.cca@meta.data$species == "Mouse"]
  hum.spg.cen[,i] <- apply(spg.exp.all.data[,hum.cells], 1, mean)
  mac.spg.cen[,i] <- apply(spg.exp.all.data[,mac.cells], 1, mean)
  mou.spg.cen[,i] <- apply(spg.exp.all.data[,mou.cells], 1, mean)
}
sum(humMacMouOrth.1to1.in$external_gene_name.x != rownames(spg.cca@data)) ##confirm gene names are ordered the same.
rownames(hum.spg.cen) <- rownames(spg.cca@data)
rownames(mac.spg.cen) <- humMacMouOrth.1to1.in$mmulatta_homolog_associated_gene_name
rownames(mou.spg.cen) <- humMacMouOrth.1to1.in$mmusculus_homolog_associated_gene_name
saveRDS(hum.spg.cen, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/hum.spg.cen.rds")
saveRDS(mac.spg.cen, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/mac.spg.cen.rds")
saveRDS(mou.spg.cen, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/mou.spg.cen.rds")
```

###. Write to xlxs
```{r}
write.xlsx2(hum.spg.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/HumanSPGClusterCentroids.xlsx", sheetName = "Human", append = FALSE, row.names = TRUE)
write.xlsx2(mac.spg.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/MonkeySPGClusterCentroids.xlsx", sheetName = "Monkey", append = FALSE, row.names = TRUE)
write.xlsx2(mou.spg.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/MouseSPGClusterCentroids.xlsx", sheetName = "Mouse", append = FALSE, row.names = TRUE)
```

Centroids for all genes for three species
extract full mouse SPG dataset
```{r}
mou.raw <- read.table("~/storage/HumanMacaqueMouseCompAnaysis/countsOriginalData/mouse_data/xaa", header = TRUE,row.names = 1,sep="\t")
#For all SPG cells
spg.cellnames <- spg.cca@cell.names[spg.cca@meta.data$species=="Mouse"]
selectcells <- colnames(mou.raw) %in% spg.cellnames
dge.mou.spg.full.raw <- mou.raw[,selectcells]
filelist <- paste0("~/storage/HumanMacaqueMouseCompAnaysis/countsOriginalData/mouse_data/", list.files("~/storage/HumanMacaqueMouseCompAnaysis/countsOriginalData/mouse_data"))
for(i in filelist[2:length(filelist)]){
print(i)
mou.raw <- read.table(i, row.names = 1, sep="\t",header = FALSE)
mou.raw <- mou.raw[,selectcells]
colnames(mou.raw) <- colnames(dge.mou.spg.full.raw)
dge.mou.spg.full.raw <- rbind(dge.mou.spg.full.raw, mou.raw)
}
rm(mou.raw)
saveRDS(dge.mou.spg.full.raw, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mou.spg.full.raw.rds")
```

```{r}
spg.cca <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/spg.cca.rds")
dge.hum.spg.full.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.hum.spg.full.raw.rds")
dge.mac.spg.full.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mac.spg.full.raw.rds")
```

```{r}
##Cluster centroid for all genes.
dge.mac.spg.full.norm <- apply(dge.mac.spg.full.raw,2,function(x) x/sum(x)*10000)
dge.hum.spg.full.norm <- apply(dge.hum.spg.full.raw,2,function(x) x/sum(x)*10000)
dge.mou.spg.full.norm <- apply(dge.mou.spg.full.raw,2,function(x) x/sum(x)*10000)
hum.spg.all.cen <- matrix(0,nrow=dim(dge.hum.spg.full.norm)[1], ncol=8)
mac.spg.all.cen <- matrix(0,nrow=dim(dge.mac.spg.full.norm)[1], ncol=8)
mou.spg.all.cen <- matrix(0,nrow=dim(dge.mou.spg.full.norm)[1], ncol=8)
#mou.spg.all.cen <- matrix(0,nrow=11029, ncol=8)
for(i in c(1:8)){
  hum.cells <- spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord == i & spg.cca@meta.data$species == "Human"]
  mac.cells <- spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord == i & spg.cca@meta.data$species == "Monkey"]
  mou.cells <- spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord == i & spg.cca@meta.data$species == "Mouse"]
  hum.spg.all.cen[,i] <- apply(dge.hum.spg.full.norm[,hum.cells], 1, mean)
  mac.spg.all.cen[,i] <- apply(dge.mac.spg.full.norm[,mac.cells], 1, mean)
  mou.spg.all.cen[,i] <- apply(dge.mou.spg.full.norm[,mou.cells], 1, mean)
}
rownames(hum.spg.all.cen) <- rownames(dge.hum.spg.full.raw)
rownames(mac.spg.all.cen) <- rownames(dge.mac.spg.full.raw)
rownames(mou.spg.all.cen) <- rownames(dge.mou.spg.full.raw)
saveRDS(hum.spg.all.cen, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata//hum.spg.all.cen.rds")
saveRDS(mac.spg.all.cen, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata//mac.spg.all.cen.rds")
saveRDS(mou.spg.all.cen, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata//mou.spg.all.cen.rds")
```

##Write to xlxs
```{r}
write.xlsx2(hum.spg.all.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/HumanSPGCluster_AllGene_Centroids.xlsx", sheetName = "Human", append = FALSE, row.names = TRUE)
write.xlsx2(mac.spg.all.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/MonkeySPGCluster_AllGene_Centroids.xlsx", sheetName = "Monkey", append = FALSE, row.names = TRUE)
write.xlsx2(mou.spg.all.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/MouseSPGCluster_AllGene_Centroids.xlsx", sheetName = "Mouse", append = FALSE, row.names = TRUE)
```

##Write to txt
```{r}
write.table(hum.spg.all.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/HumanSPGCluster_AllGene_Centroids.txt", quote = FALSE, row.names = TRUE, sep="\t")
write.table(mac.spg.all.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/MonkeySPGCluster_AllGene_Centroids.txt", quote = FALSE, row.names = TRUE, sep="\t")
write.table(mou.spg.all.cen[,c(1:6)], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_centroids/MouseSPGCluster_AllGene_Centroids.txt", quote = FALSE, row.names = TRUE, sep="\t")
```

Centroid rank correlation for three species
```{r fig.width=5, fig.height=4.5}
hum.spg.cor <- cor(hum.spg.all.cen[,c(1:6)], method="spearman")
Heatmap(hum.spg.cor, cluster_rows = FALSE, cluster_columns = FALSE)
```
```{r fig.width=5, fig.height=4.5}
mac.spg.cor <- cor(mac.spg.all.cen[,c(1:6)], method="spearman")
Heatmap(mac.spg.cor, cluster_rows = FALSE, cluster_columns = FALSE)
```
```{r fig.width=5, fig.height=4.5}
mou.spg.cor <- cor(mou.spg.all.cen[,c(1:6)], method="spearman")
Heatmap(mou.spg.cor, cluster_rows = FALSE, cluster_columns = FALSE)
```

##. Plot all the clusters in three species
```{r fig.width=20, fig.height=9}
plist <- list()
for(i in c(1:8)){
  spg.cca@meta.data$var <- (spg.cca@meta.data$res.0.6.ord==i & spg.cca@meta.data$species=="Human")*i
  p <- TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "var", colors.use = c("grey",cols[i]), plot.order = c(i,"0"))
  plist[[i]] = p
}
do.call("grid.arrange", c(plist, ncol=4))
```
```{r fig.width=20, fig.height=9}
plist <- list()
for(i in c(1:8)){
  spg.cca@meta.data$var <- (spg.cca@meta.data$res.0.6.ord==i & spg.cca@meta.data$species=="Monkey")*i
  p <- TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "var", colors.use = c("grey",cols[i]), plot.order = c(i,"0"))
  plist[[i]] = p
}
do.call("grid.arrange", c(plist, ncol=4))
```




```{r fig.width=20, fig.height=9}
plist <- list()
for(i in c(1:8)){
  spg.cca@meta.data$var <- (spg.cca@meta.data$res.0.6.ord==i & spg.cca@meta.data$species=="Mouse")*i
  p <- TSNEPlot(object = spg.cca, do.return = TRUE, pt.size = 1, group.by = "var", colors.use = c("grey",cols[i]), plot.order = c(i,"0"))
  plist[[i]] = p
}
do.call("grid.arrange", c(plist, ncol=4))
```

##. Markers for clusters 1 and 2 in mouse
```{r}
spg.cca@meta.data$speclus <- paste0(spg.cca@meta.data$species,"_", spg.cca@meta.data$res.0.6.ord)
spg.cca <- SetAllIdent(spg.cca, id = "speclus")
marker1over2 <- FindMarkers(spg.cca, ident.1 = "Mouse_1", ident.2 = "Mouse_2", logfc.threshold = 0.6, min.diff.pct = 0.1)
```

```{r}
rownames(humMacMouOrth.1to1.in) = humMacMouOrth.1to1.in$external_gene_name.x
rownames(marker1over2) <- humMacMouOrth.1to1.in[rownames(marker1over2),]$mmusculus_homolog_associated_gene_name
```

```{r}
write.table(marker1over2, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/MakersMouseClust1vs2.txt",quote = FALSE, row.names = TRUE, col.names = TRUE, sep = "\t")
```

##Call markers using all genes for each species
###Human

spg.raw file used below are files that removed cluster 7-8 in SPG CCA clusterign 
```{r}
dge.hum.spg <- CreateSeuratObject(raw.data = dge.hum.spg.raw[,spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6) & spg.cca@meta.data$species=="Human"]])
dge.hum.spg <- NormalizeData(object = dge.hum.spg, normalization.method = "LogNormalize", 
    scale.factor = 10000)
dge.hum.spg <- ScaleData(object = dge.hum.spg,vars.to.regress = c("nUMI"))
```

```{r}
dge.hum.spg@meta.data$clus <- spg.cca@meta.data[dge.hum.spg@cell.names,]$res.0.6.ord
dge.hum.spg <- SetIdent(dge.hum.spg, ident.use = dge.hum.spg@meta.data$clus)
```

```{r}
hum.spg.markers <- list()
for(i in c(1:6)){
  hum.spg.markers[[i]] <- FindMarkers(dge.hum.spg, ident.1 = i, logfc.threshold = 0.25, min.pct = 0.1, only.pos = TRUE)
}
```

```{r}
for(i in c(1:6)){
if(i==1){
  write.xlsx2(x = hum.spg.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_hum_6Clusters_Markers_allGene_lgfc0.25pct10.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = hum.spg.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_hum_6Clusters_Markers_allGene_lgfc0.25pct10.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```

```{r}
dge.mac.spg <- CreateSeuratObject(raw.data = dge.mac.spg.raw[,spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6) & spg.cca@meta.data$species=="Monkey"]])
dge.mac.spg <- NormalizeData(object = dge.mac.spg, normalization.method = "LogNormalize", 
    scale.factor = 10000)
dge.mac.spg <- ScaleData(object = dge.mac.spg,vars.to.regress = c("nUMI"))

dge.mac.spg@meta.data$clus <- spg.cca@meta.data[dge.mac.spg@cell.names,]$res.0.6.ord
dge.mac.spg <- SetIdent(dge.mac.spg, ident.use = dge.mac.spg@meta.data$clus)

mac.spg.markers <- list()
for(i in c(1:6)){
  mac.spg.markers[[i]] <- FindMarkers(dge.mac.spg, ident.1 = i, logfc.threshold = 0.25, min.pct = 0.1, only.pos = TRUE)
}
```

```{r}
for(i in c(1:6)){
if(i==1){
  write.xlsx2(x = mac.spg.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_mac_6Clusters_Markers_allGene_lgfc0.25pct10.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = mac.spg.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_mac_6Clusters_Markers_allGene_lgfc0.25pct10.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```

```{r}
dge.mou.spg <- CreateSeuratObject(raw.data = dge.mou.spg.raw[,spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6) & spg.cca@meta.data$species=="Mouse"]])
dge.mou.spg <- NormalizeData(object = dge.mou.spg, normalization.method = "LogNormalize", 
    scale.factor = 10000)
dge.mou.spg <- ScaleData(object = dge.mou.spg,vars.to.regress = c("nUMI"))

dge.mou.spg@meta.data$clus <- spg.cca@meta.data[dge.mou.spg@cell.names,]$res.0.6.ord
dge.mou.spg <- SetIdent(dge.mou.spg, ident.use = dge.mou.spg@meta.data$clus)

mou.spg.markers <- list()
for(i in c(1:6)){
  mou.spg.markers[[i]] <- FindMarkers(dge.mou.spg, ident.1 = i, logfc.threshold = 0.25, min.pct = 0.1, only.pos = TRUE)
}
```

```{r}
for(i in c(1:6)){
if(i==1){
  write.xlsx2(x = mou.spg.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_mou_6Clusters_Markers_allGene_lgfc0.25pct10.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = mou.spg.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_mou_6Clusters_Markers_allGene_lgfc0.25pct10.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```

```{r}
hum.spg.markers.df <- do.call("rbind", hum.spg.markers)
mac.spg.markers.df <- do.call("rbind", mac.spg.markers)
mou.spg.markers.df <- do.call("rbind", mou.spg.markers)
hum.spg.stages <- c()
mac.spg.stages <- c()
mou.spg.stages <- c()
for(i in c(1:6)){
  hum.spg.stages <- c(hum.spg.stages, rep(i, dim(hum.spg.markers[[i]])[1]))
  mac.spg.stages <- c(mac.spg.stages, rep(i, dim(mac.spg.markers[[i]])[1]))
  mou.spg.stages <- c(mou.spg.stages, rep(i, dim(mou.spg.markers[[i]])[1]))
}
hum.spg.markers.df$stages <- hum.spg.stages
mac.spg.markers.df$stages <- mac.spg.stages
mou.spg.markers.df$stages <- mou.spg.stages
```

```{r}
hum.spg.markers.df <- hum.spg.markers.df[hum.spg.markers.df$p_val_adj < 0.01,]
mac.spg.markers.df <- mac.spg.markers.df[mac.spg.markers.df$p_val_adj < 0.01,]
mou.spg.markers.df <- mou.spg.markers.df[mou.spg.markers.df$p_val_adj < 0.01,]
```

```{r}
hum.spg.markers.df$gene <- rownames(hum.spg.markers.df)
mac.spg.markers.df$gene <- rownames(mac.spg.markers.df)
mou.spg.markers.df$gene <- rownames(mou.spg.markers.df)
```

```{r}
hum.spg.marker.orth <- hum.spg.markers.df[hum.spg.markers.df$gene %in% humMacMouOrth.1to1.in$external_gene_name.x,]
hum.spg.marker.orth$humorth <- hum.spg.marker.orth$gene

mac.spg.marker.orth <- mac.spg.markers.df[mac.spg.markers.df$gene %in% humMacMouOrth.1to1.in$mmulatta_homolog_associated_gene_name,]
rownames(humMacMouOrth.1to1.in) = humMacMouOrth.1to1.in$mmulatta_homolog_associated_gene_name
mac.spg.marker.orth$humorth <- humMacMouOrth.1to1.in[mac.spg.marker.orth$gene,]$external_gene_name.x

mou.spg.marker.orth <- mou.spg.markers.df[mou.spg.markers.df$gene %in% humMacMouOrth.1to1.in$mmusculus_homolog_associated_gene_name,]
rownames(humMacMouOrth.1to1.in) = humMacMouOrth.1to1.in$mmusculus_homolog_associated_gene_name
mou.spg.marker.orth$humorth <- humMacMouOrth.1to1.in[mou.spg.marker.orth$gene,]$external_gene_name.x
```


hummac hum
```{r}
spg.hum.mac.hum.clus.num <- c()
for(i in c(1:6)){
  a = length(intersect(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i], mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i]))
  b = length(intersect(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i], mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages!=i]))
  c = length(setdiff(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i], mac.spg.marker.orth$humorth))
  d = length(hum.spg.markers.df$gene[hum.spg.markers.df$stages==i])-length(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i])
  spg.hum.mac.hum.clus.num <- cbind(spg.hum.mac.hum.clus.num,c(a,b,c,d))
}
spg.hum.mac.hum.clus.num
```

hummac mac
```{r}
spg.hum.mac.mac.clus.num <- c()
for(i in c(1:6)){
  a = length(intersect(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i], hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i]))
  b = length(intersect(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i], hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages!=i]))
  c = length(setdiff(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i], hum.spg.marker.orth$humorth))
  d = length(mac.spg.markers.df$gene[mac.spg.markers.df$stages==i])-length(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i])
  spg.hum.mac.mac.clus.num <- cbind(spg.hum.mac.mac.clus.num,c(a,b,c,d))
}
spg.hum.mac.mac.clus.num
```

```{r}
hummac.spg.rat.3 <- apply(spg.hum.mac.hum.clus.num,2,function(x) x/sum(x[c(1:3)]))[c(1:3),]/2 + apply(spg.hum.mac.mac.clus.num,2,function(x) x/sum(x[c(1:3)]))[c(1:3),]/2
hummac.spg.rat.4 <- apply(spg.hum.mac.hum.clus.num,2,function(x) x/sum(x))/2 + apply(spg.hum.mac.mac.clus.num,2,function(x) x/sum(x))/2
hummac.spg.rat.3
hummac.spg.rat.4
```


hummou hum
```{r}
spg.hum.mou.hum.clus.num <- c()
for(i in c(1:6)){
  a = length(intersect(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i], mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i]))
  b = length(intersect(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i], mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages!=i]))
  c = length(setdiff(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i], mou.spg.marker.orth$humorth))
  d = length(hum.spg.markers.df$gene[hum.spg.markers.df$stages==i])-length(hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i])
  spg.hum.mou.hum.clus.num <- cbind(spg.hum.mou.hum.clus.num,c(a,b,c,d))
}
spg.hum.mou.hum.clus.num
```

hummou mou
```{r}
spg.hum.mou.mou.clus.num <- c()
for(i in c(1:6)){
  a = length(intersect(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i], hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages==i]))
  b = length(intersect(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i], hum.spg.marker.orth$humorth[hum.spg.marker.orth$stages!=i]))
  c = length(setdiff(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i], hum.spg.marker.orth$humorth))
  d = length(mou.spg.markers.df$gene[mou.spg.markers.df$stages==i])-length(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i])
  spg.hum.mou.mou.clus.num <- cbind(spg.hum.mou.mou.clus.num,c(a,b,c,d))
}
spg.hum.mou.mou.clus.num
```

```{r}
hummou.spg.rat.3 <- apply(spg.hum.mou.hum.clus.num,2,function(x) x/sum(x[c(1:3)]))[c(1:3),]/2 + apply(spg.hum.mou.mou.clus.num,2,function(x) x/sum(x[c(1:3)]))[c(1:3),]/2
hummou.spg.rat.4 <- apply(spg.hum.mou.hum.clus.num,2,function(x) x/sum(x))/2 + apply(spg.hum.mou.mou.clus.num,2,function(x) x/sum(x))/2
hummou.spg.rat.3
hummou.spg.rat.4
```

macmou mac
```{r}
spg.mac.mou.mac.clus.num <- c()
for(i in c(1:6)){
  a = length(intersect(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i], mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i]))
  b = length(intersect(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i], mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages!=i]))
  c = length(setdiff(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i], mou.spg.marker.orth$humorth))
  d = length(mac.spg.markers.df$gene[mac.spg.markers.df$stages==i])-length(mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i])
  spg.mac.mou.mac.clus.num <- cbind(spg.mac.mou.mac.clus.num,c(a,b,c,d))
}
spg.mac.mou.mac.clus.num
```

macmou mou
```{r}
spg.mac.mou.mou.clus.num <- c()
for(i in c(1:6)){
  a = length(intersect(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i], mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages==i]))
  b = length(intersect(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i], mac.spg.marker.orth$humorth[mac.spg.marker.orth$stages!=i]))
  c = length(setdiff(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i], mac.spg.marker.orth$humorth))
  d = length(mou.spg.markers.df$gene[mou.spg.markers.df$stages==i])-length(mou.spg.marker.orth$humorth[mou.spg.marker.orth$stages==i])
  spg.mac.mou.mou.clus.num <- cbind(spg.mac.mou.mou.clus.num,c(a,b,c,d))
}
spg.mac.mou.mou.clus.num
```

```{r}
macmou.spg.rat.3 <- apply(spg.mac.mou.mac.clus.num,2,function(x) x/sum(x[c(1:3)]))[c(1:3),]/2 + apply(spg.mac.mou.mou.clus.num,2,function(x) x/sum(x[c(1:3)]))[c(1:3),]/2
macmou.spg.rat.4 <- apply(spg.mac.mou.mac.clus.num,2,function(x) x/sum(x))/2 + apply(spg.mac.mou.mou.clus.num,2,function(x) x/sum(x))/2
macmou.spg.rat.3
macmou.spg.rat.4
```


```{r}
hummac.spg.rat.3
hummou.spg.rat.3
macmou.spg.rat.3
```

```{r}
hummac.spg.rat.4
hummou.spg.rat.4
macmou.spg.rat.4
```
```{r}
stakbar3 <- function(dat){
colnames(dat) <- c(1:6)
rownames(dat) <- c("In Phase", "Shifted", "Unique")
dat <- data.frame(t(dat))
dat$Cluster <- c(1:6)
df <- melt(dat, id.var="Cluster")
p <- ggplot(df, aes(x = Cluster, y = value, fill = variable)) + 
  geom_bar(stat = "identity",color = "grey40",width = 1, alpha=0.8) + 
  scale_x_continuous(breaks=c(1:6)) + 
  theme(axis.title.y=element_blank(),
        legend.title=element_blank()) + 
  scale_fill_manual(values = c("#323232","#767676","#cecece"))
print(p)
}
```

```{r fig.width=5, fig.height=4.5}
stakbar3(hummac.spg.rat.3)
stakbar3(hummou.spg.rat.3)
stakbar3(macmou.spg.rat.3)
```

```{r}
stakbar4 <- function(dat){
colnames(dat) <- c(1:6)
rownames(dat) <- c("In Phase", "Shifted", "Unique (1-1-1 Ortholog)", "Unique (Non 1-1-1 Ortholog)")
dat <- data.frame(t(dat))
dat$Cluster <- c(1:6)
df <- melt(dat, id.var="Cluster")
p <- ggplot(df, aes(x = Cluster, y = value, fill = variable)) + 
  geom_bar(stat = "identity",color = "grey40",width = 1, alpha=0.8) + 
  scale_x_continuous(breaks=c(1:6)) + 
  theme(axis.title.y=element_blank(),
        legend.title=element_blank()) + 
  scale_fill_manual(values = c("#5a5a5a","#737373","#8d8d8d","#a6a6a6"))
print(p)
}
```

```{r fig.width=6, fig.height=4.5}
stakbar4(hummac.spg.rat.4)
stakbar4(hummou.spg.rat.4)
stakbar4(macmou.spg.rat.4)
```

Plot markers
```{r}
markers <- c("THY1", "SALL4", "PLZF", "ZBTB16", "ID4", "GFRA1", "ITGA6", "EPCAM", "PAX7", "BMI1", "EOMES", "UTF1", "UCHL1", "LIN28A", "LIN28B", "PIWIL4", "MORC1", "CDK17", "TSPAN33", "MAGEB2", "ITGA6", "MKI67", "L1TD1", "ID1", "ID3", "STRA8", "GPC3", "GPC4", "GPX1", "TCF3", "FMR1", "DNAJB6", "LENG8", "C19orf84", "ZNF428", "PSIP1", "SSFA2", "APOE", "CRYM")
```


```{r}
humMacMouOrth.1to1.in <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/humMacMouOrth.1to1.in.Rds")
spg.cca <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/spg.cca.rds")
```




```{r}
redblue100.alpha<-rgb(read.table("~/storage/HumanMacaqueMouseCompAnaysis/redblue100.txt",sep='\t',row.names=1,header=T),alpha=0.8)

set.ifnull=function(x,y) {
  if(is.null(x)) return(y)
  return(x)
}
translate.dim.code=function(reduction.use) {
  if (reduction.use=="pca") return.code="PC"
  if (reduction.use=="ica") return.code="IC"
  if (reduction.use=="tsne") return.code="tSNE_"
  if (reduction.use=="mds") return.code="MDS"
  if (reduction.use=="umap") return.code="UMAP"
  return(return.code)
}

plotgene <- function(gene){
feature=features.plot=gene

dim.1 = 1; dim.2 = 2; cells.use = NULL; pt.size = 1;
pch.use = 16; reduction.use = "tsne";
use.imputed = FALSE; no.axes = TRUE; no.legend = FALSE

cells.use <- set.ifnull(cells.use, spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dim.code <- translate.dim.code(reduction.use)
dim.codes <- paste(dim.code, c(dim.1, dim.2), sep = "")
data.plot <- as.data.frame(spg.cca@dr$tsne@cell.embeddings[cells.use,])

x1 <- paste(dim.code, dim.1, sep = "")
x2 <- paste(dim.code, dim.2, sep = "")

data.plot$x <- data.plot[cells.use, x1]
data.plot$y <- data.plot[cells.use, x2]
data.plot$pt.size <- pt.size
data.use <- data.frame(t(spg.cca@data[feature,cells.use]))
rownames(data.use)=feature

data.gene0 <- na.omit(data.frame(data.use[feature, ]))
data.plot$gene <- t(data.gene0)

st6<- data.plot
st6<-st6[order(st6[,6]),]
z<-st6[,6]

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/plots/newplots/spg.cca","_",feature,"_40-100redblue.jpeg"),height=850,width=800,res=300)
par(mar=c(0.5,0.5,2,0.5),mgp=c(1,0.5,0))
z<-st6[,6]
zcolor <- redblue100.alpha[40:100][(z - min(z))/diff(range(z))*60 + 1]
plot(st6[,1],st6[,2], col=zcolor,pch=19,cex=0.3,cex.main=1.5,axes=F,main=features.plot,xlab="",ylab="")
dev.off()
}
```

```{r}
for(i in intersect(markers, humMacMouOrth.1to1.in$external_gene_name.x)){
  plotgene(i)
}
```

```{r}
markers2 <- c("RARA","RARB","RARG","RXRA","RXRB","ALDH1A1","ALDH1A2","ALDH1A3")
for(i in intersect(markers2, humMacMouOrth.1to1.in$external_gene_name.x)){
  plotgene(i)
}
```
"RARA"    "RARB"    "RARG"    "RXRA"    "RXRB"    "ALDH1A2" "ALDH1A3" added to the three species SPG

```{r}
marker3 <- c("TSPAN33", "PIWIL4", "MORC1", "TCF3", "CDK17", "MAGEB2", "DNAJB6", "FMR1", "ZBTB43", "ZNF428", "MSL3", "LY6K","VCX","VCY")
for(i in intersect(marker3, humMacMouOrth.1to1.in$external_gene_name.x)){
  plotgene(i)
}
```

```{r}
marker5 <- c("HORMAD1","DNMT1","MEIOB","SOHLH1","SOHLH2","KIT","KIF1A","TEX101","SYCP2","FMR1NB","DMRT1","DMRTB1")
for(i in intersect(marker5, humMacMouOrth.1to1.in$external_gene_name.x)){
  plotgene(i)
}
```

```{r}
marker6 <- c("GPC3","GPX1")
```

```{r}
plotgene("GPC3")
```

```{r}
plotgene("DMRT1")
```
```{r}
markers <- c("TSPAN33","PIWIL4","MORC1","TCF3","CDK17","STRA8","MAGEB2","FMR1","DNAJB6","L1TD1")
```


```{r fig.width=10,fig.height=20}
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/AlignedThreeSpecies_Violin.pdf",width = 10, height = 20)
VlnPlot(object = spg.cca, features.plot = intersect(markers,humMacMouOrth.1to1.in$external_gene_name.x),nCol = 4,group.by = "res.0.6.ord",point.size.use=0, ident.include = c(1:6))
dev.off()
```

```{r}
for(i in intersect(marker5, humMacMouOrth.1to1.in$external_gene_name.x)){
  pdf(paste0("/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/",i,".pdf"),width = 2.5, height = 2.5)
  p<-VlnPlot(object = spg.cca, features.plot = i,nCol = 4,group.by = "res.0.6.ord",point.size.use=0, ident.include = c(1:6),cells.use=spg.cca@cell.names[spg.cca@meta.data$species=="Human"])
  print(p)
  dev.off()
}
```

```{r}
spg.cca.hum <- SubsetData(object = spg.cca, cells.use = spg.cca@cell.names[spg.cca@meta.data$species=="Human"])
spg.cca.mac <- SubsetData(object = spg.cca, cells.use = spg.cca@cell.names[spg.cca@meta.data$species=="Monkey"])
spg.cca.mou <- SubsetData(object = spg.cca, cells.use = spg.cca@cell.names[spg.cca@meta.data$species=="Mouse"])
```

```{r}
for(i in intersect(markers, humMacMouOrth.1to1.in$external_gene_name.x)){
  jpeg(paste0("/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/spg.cca","_",i,".human.violin.jpeg"),height=850,width=800,res=300)
  p<-VlnPlot(object = spg.cca.hum, features.plot = i,group.by = "res.0.6.ord",point.size.use=0, ident.include = c(1:6),cols.use = cols)
  print(p)
  dev.off()
  
  jpeg(paste0("/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/spg.cca","_",i,".monkey.violin.jpeg"),height=850,width=800,res=300)
  p<-VlnPlot(object = spg.cca.mac, features.plot = i,group.by = "res.0.6.ord",point.size.use=0, ident.include = c(1:6),cols.use = cols)
  print(p)
  dev.off()
  
  jpeg(paste0("/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/spg.cca","_",i,".mouse.violin.jpeg"),height=850,width=800,res=300)
  p<-VlnPlot(object = spg.cca.mou, features.plot = i,group.by = "res.0.6.ord",point.size.use=0, ident.include = c(1:6),cols.use = cols)
  print(p)
  dev.off()
}
```


```{r}
dge.hum.spg.norm <- apply(dge.hum.spg.raw,2,function(x) x/sum(x)*1e4)
dge.mac.spg.norm <- apply(dge.mac.spg.raw,2,function(x) x/sum(x)*1e4)
dge.mou.spg.norm <- apply(dge.mou.spg.raw,2,function(x) x/sum(x)*1e4)
```


```{r}
plts <- list()
j=0
for(i in intersect(markers, rownames(dge.hum.spg.norm))){
  j = j+1
  dat = data.frame(expr = log(dge.hum.spg.norm[i,]+1), ident = spg.cca@ident[colnames(dge.hum.spg.norm)])
  plts[[j]] <- ggplot(dat, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle(i) + theme(legend.position = "none")
}
```

```{r}
library(gridExtra)
```

```{r fig.width=10, fig.height=30}
pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/Hum_Violin.pdf",width = 10, height = 30)
do.call("grid.arrange", c(plts, ncol=4))
dev.off()
```

```{r}
plts <- list()
j=0
#mac.markers <- humMacMouOrth.1to1.in$mmulatta_homolog_associated_gene_name[humMacMouOrth.1to1.in$external_gene_name.x %in% markers]
for(i in intersect(markers, rownames(dge.mac.spg.norm))){
  j = j+1
  dat = data.frame(expr = log(dge.mac.spg.norm[i,]+1), ident = spg.cca@ident[colnames(dge.mac.spg.norm)])
  plts[[j]] <- ggplot(dat, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle(i) + theme(legend.position = "none")
}

pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/Mac_Violin.pdf",width = 10, height = 30)
do.call("grid.arrange", c(plts, ncol=4))
dev.off()
```

```{r}
plts <- list()
j=0
mou.markers <- humMacMouOrth.1to1.in$mmusculus_homolog_associated_gene_name[humMacMouOrth.1to1.in$external_gene_name.x %in% markers]
for(i in mou.markers){
  j = j+1
  dat = data.frame(expr = log(dge.mou.spg.norm[i,]+1), ident = spg.cca@ident[colnames(dge.mou.spg.norm)])
  plts[[j]] <- ggplot(dat, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle(i) + theme(legend.position = "none")
}

pdf("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/Mou_Violin.pdf",width = 10, height = 30)
do.call("grid.arrange", c(plts, ncol=4))
dev.off()
```

```{r}
humMac[humMac$external_gene_name %in% marker4, c(1:4)]
humMou[humMou$external_gene_name %in% marker4, c(1:4)]
```

```{r}
marker5.hum <- c("HORMAD1","DNMT1","MEIOB","SOHLH1","SOHLH2","KIT","KIF1A","TEX101","SYCP2","FMR1NB","VCX", "VCX2", "VCX3A", "VCX3B")
marker5.mac <- c("HORMAD1","DNMT1", "MEIOB", "KIF1A",  "TEX101","SYCP2",  "FMR1NB", "ENSMMUG00000001172")
# "ENSMMUG00000001172" is ortholog for SOHLH2
marker5.mou <- c("Sycp2","Tex101","Sohlh2","Sohlh1","Kif1a","Fmr1nb","Hormad1", "Meiob","Dnmt1","Kit")
```

```{r}
i="GPC3"
  pdf(paste0("/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/",i,".pdf"),width = 2.5, height = 2.5)
  p<-VlnPlot(object = spg.cca, features.plot = i,nCol = 4,group.by = "res.0.6.ord",point.size.use=0, ident.include = c(1:6))
  print(p)
  dev.off()
```

```{r}
marker5.hum <- c("DMRT1","DMRTB1")
marker5.mac <- c("DMRT1")
marker5.mou <- c("Dmrt1","Dmrtb1")
```

```{r}
marker5.hum <- marker6
marker5.mac <- c("GPC3")
marker5.mou <- c("Gpc3","Gpx1")
marker6.mac <- c("ENSMMUG00000046487", "ENSMMUG00000041391")
```


```{r}
dge.hum.spg.full.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.hum.spg.full.raw.rds")
dge.mac.spg.full.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mac.spg.full.raw.rds")
dge.mou.spg.full.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mou.spg.full.raw.rds")
```

```{r}
dge.hum.spg.full.norm <- apply(dge.hum.spg.full.raw,2,function(x) x/sum(x)*1e4)
dge.mac.spg.full.norm <- apply(dge.mac.spg.full.raw,2,function(x) x/sum(x)*1e4)
dge.mou.spg.full.norm <- apply(dge.mou.spg.full.raw,2,function(x) x/sum(x)*1e4)
```

```{r}
for(i in intersect(marker5.mou, rownames(dge.mou.spg.full.norm))){
  dat = data.frame(expr = log(dge.mou.spg.full.norm[i,]+1), ident = spg.cca@ident[colnames(dge.mou.spg.full.norm)])
  dat = dat[dat$ident %in% c(1:6), ]
  p <- ggplot(dat, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle(i) + theme(legend.position = "none")
  pdf(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Mou_",i,".pdf"),width = 2.5, height = 2.5)
  print(p)
  dev.off()
}

```

```{r}
for(i in intersect(marker5.hum, rownames(dge.hum.spg.full.norm))){
  dat = data.frame(expr = log(dge.hum.spg.full.norm[i,]+1), ident = spg.cca@ident[colnames(dge.hum.spg.full.norm)])
  dat = dat[dat$ident %in% c(1:6), ]
  p <- ggplot(dat, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle(i) + theme(legend.position = "none")
  pdf(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Hum_",i,".pdf"),width = 2.5, height = 2.5)
  print(p)
  dev.off()
}

```

```{r}
for(i in intersect(marker6.mac, rownames(dge.mac.spg.full.norm))){
  dat = data.frame(expr = log(dge.mac.spg.full.norm[i,]+1), ident = spg.cca@ident[colnames(dge.mac.spg.full.norm)])
  dat = dat[dat$ident %in% c(1:6), ]
  p <- ggplot(dat, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle(i) + theme(legend.position = "none")
  pdf(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Mac_",i,".pdf"),width = 3, height = 2.5)
  print(p)
  dev.off()
}

```


1. Read in datasets (3-12)
2. Run Seurat CCA for SPG from three species (14-88)
3. Do clustering and plot the clusters (90-158)
4. Compare with the mouse SPG clusters in previous paper (163-187)
5. Markers finding and markers plots (based on 1-1-1 orthologues) (190-241)
6. Calculate SPG cluster centroid for each species (242-352)
7. Call Markers using all genes for each species separately (400-479)
8. Plot extra markers, polished figures on tSNE and violin plots (481-621)

Plot #Genes and #UMIs for 6 clusters.
```{r}
#Human
hum.genes <- apply(dge.hum.spg.full.raw,2, function(x) sum(x!=0))
hum.umis <- apply(dge.hum.spg.full.raw,2,sum)
df <- data.frame(NumGene = hum.genes, NumUMIs = hum.umis, Cluster=spg.cca@meta.data[colnames(dge.hum.spg.full.raw),]$res.0.6.ord)
df <- df[df$Cluster %in% c(1:6),]
p1 <- ggplot(df, aes(x=Cluster,y=NumGene, fill=Cluster))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("#Genes (Human)") + theme(legend.position = "none")
p2 <- ggplot(df, aes(x=Cluster,y=NumUMIs, fill=Cluster))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("#UMIs (Human)") + theme(legend.position = "none")
pdf(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/spg.hum.NumGeneUMI.violin.pdf"),width = 6, height = 2.5)
do.call("grid.arrange", c(list(p1,p2), ncol=2))
dev.off()
```

```{r}
#Monkey
mac.genes <- apply(dge.mac.spg.full.raw,2, function(x) sum(x!=0))
mac.umis <- apply(dge.mac.spg.full.raw,2,sum)
df <- data.frame(NumGene = mac.genes, NumUMIs = mac.umis, Cluster=spg.cca@meta.data[colnames(dge.mac.spg.full.raw),]$res.0.6.ord)
df <- df[df$Cluster %in% c(1:6),]
p1 <- ggplot(df, aes(x=Cluster,y=NumGene, fill=Cluster))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("#Genes (Monkey)") + theme(legend.position = "none")
p2 <- ggplot(df, aes(x=Cluster,y=NumUMIs, fill=Cluster))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("#UMIs (Monkey)") + theme(legend.position = "none")
pdf(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/spg.mac.NumGeneUMI.violin.pdf"),width = 6, height = 2.5)
do.call("grid.arrange", c(list(p1,p2), ncol=2))
dev.off()
```

```{r}
#Mouse
mou.genes <- apply(dge.mou.spg.full.raw,2, function(x) sum(x!=0))
mou.umis <- apply(dge.mou.spg.full.raw,2,sum)
df <- data.frame(NumGene = mou.genes, NumUMIs = mou.umis, Cluster=spg.cca@meta.data[colnames(dge.mou.spg.full.raw),]$res.0.6.ord)
df <- df[df$Cluster %in% c(1:6),]
p1 <- ggplot(df, aes(x=Cluster,y=NumGene, fill=Cluster))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("#Genes (Mouse)") + theme(legend.position = "none")
p2 <- ggplot(df, aes(x=Cluster,y=NumUMIs, fill=Cluster))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("#UMIs (Mouse)") + theme(legend.position = "none")
pdf(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/spg.mou.NumGeneUMI.violin.pdf"),width = 6, height = 2.5)
do.call("grid.arrange", c(list(p1,p2), ncol=2))
dev.off()
```

```{r}
plotgene_spe <- function(gene, cells.use=NULL){
feature=features.plot=gene

dim.1 = 1; dim.2 = 2; pt.size = 1;
pch.use = 16; reduction.use = "tsne";
use.imputed = FALSE; no.axes = TRUE; no.legend = FALSE

cells.use <- set.ifnull(cells.use, spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dim.code <- translate.dim.code(reduction.use)
dim.codes <- paste(dim.code, c(dim.1, dim.2), sep = "")
data.plot <- as.data.frame(spg.cca@dr$tsne@cell.embeddings[cells.use,])

x1 <- paste(dim.code, dim.1, sep = "")
x2 <- paste(dim.code, dim.2, sep = "")

data.plot$x <- data.plot[cells.use, x1]
data.plot$y <- data.plot[cells.use, x2]
data.plot$pt.size <- pt.size
data.use <- data.frame(t(spg.cca@data[feature,cells.use]))
rownames(data.use)=feature

data.gene0 <- na.omit(data.frame(data.use[feature, ]))
data.plot$gene <- t(data.gene0)

st6<- data.plot
st6<-st6[order(st6[,6]),]
z<-st6[,6]

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/plots/newplots/spg.cca","_",feature,"_40-100redblue.jpeg"),height=850,width=800,res=300)
par(mar=c(0.5,0.5,2,0.5),mgp=c(1,0.5,0))
z<-st6[,6]
zcolor <- redblue100.alpha[40:100][(z - min(z))/diff(range(z))*60 + 1]
plot(st6[,1],st6[,2], col=zcolor,pch=19,cex=0.3,cex.main=1.5,axes=F,main=features.plot,xlab="",ylab="")
dev.off()
}
```


```{r}
plotfeature_spe <- function(gene, cells.use=NULL){
feature=features.plot=gene

dim.1 = 1; dim.2 = 2; pt.size = 1;
pch.use = 16; reduction.use = "tsne";
use.imputed = FALSE; no.axes = TRUE; no.legend = FALSE

cells.use <- set.ifnull(cells.use, spg.cca@cell.names[spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
dim.code <- translate.dim.code(reduction.use)
dim.codes <- paste(dim.code, c(dim.1, dim.2), sep = "")
data.plot <- as.data.frame(spg.cca@dr$tsne@cell.embeddings[cells.use,])

x1 <- paste(dim.code, dim.1, sep = "")
x2 <- paste(dim.code, dim.2, sep = "")

data.plot$x <- data.plot[cells.use, x1]
data.plot$y <- data.plot[cells.use, x2]
data.plot$pt.size <- pt.size
data.use <- data.frame(t(spg.cca@meta.data[cells.use,feature]))
rownames(data.use)=feature

data.gene0 <- na.omit(data.frame(data.use[feature, ]))
data.plot$gene <- t(data.gene0)

st6<- data.plot
st6<-st6[order(st6[,6]),]
z<-st6[,6]

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/plots/newplots/spg.cca","_",feature,"_40-100redblue.jpeg"),height=850,width=800,res=300)
par(mar=c(0.5,0.5,2,0.5),mgp=c(1,0.5,0))
z<-st6[,6]
zcolor <- redblue100.alpha[40:100][(z - min(z))/diff(range(z))*60 + 1]
plot(st6[,1],st6[,2], col=zcolor,pch=19,cex=0.3,cex.main=1.5,axes=F,main=features.plot,xlab="",ylab="")
dev.off()
}
```

```{r}
spg.cca@meta.data$VCX <- 0
spg.cca@meta.data[colnames(dge.hum.spg.full@data),]$VCX <- dge.hum.spg.full@data["VCX",]
plotfeature_spe("VCX")
```

```{r}
spg.cca@meta.data$VCX2 <- 0
spg.cca@meta.data[colnames(dge.hum.spg.full@data),]$VCX2 <- dge.hum.spg.full@data["VCX2",]
plotfeature_spe("VCX2")
```
```{r}
spg.cca@meta.data$VCX3A <- 0
spg.cca@meta.data[colnames(dge.hum.spg.full@data),]$VCX3A <- dge.hum.spg.full@data["VCX3A",]
plotfeature_spe("VCX3A")
```

```{r}
spg.cca@meta.data$VCX3B <- 0
spg.cca@meta.data[colnames(dge.hum.spg.full@data),]$VCX3B <- dge.hum.spg.full@data["VCX3B",]
plotfeature_spe("VCX3B")
```

```{r}
plotgene_spe("NANOS3", cells.use = spg.cca@cell.names[spg.cca@meta.data$species=="Mouse" & spg.cca@meta.data$res.0.6.ord %in% c(1:6)])
```


```{r fig.width=3, fig.height=2.5}
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
df <- data.frame(expr = spg.cca.hum@data["NANOS3",], ident = spg.cca.hum@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width", adjust=2) + ylab("") + ggtitle("NANOS3")  + scale_fill_manual(values = cols)+ theme(legend.position = "none")
```

```{r fig.width=3, fig.height=2.5}
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
df <- data.frame(expr = spg.cca.mac@data["NANOS3",], ident = spg.cca.mac@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width", adjust=2) + xlab("Identity") + ylab("") + ggtitle("NANOS3")  + scale_fill_manual(values = cols)+ theme(legend.position = "none")
```


```{r fig.width=3, fig.height=2.5}
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
df <- data.frame(expr = spg.cca.mou@data["NANOS3",], ident = spg.cca.mou@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width", adjust=2) + xlab("Identity") + ylab("") + ggtitle("NANOS3")  + scale_fill_manual(values = cols)+ theme(legend.position = "none")
```

```{r}
df <- data.frame(expr = spg.cca.mou@data["STRA8",], ident = spg.cca.mou@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width") + xlab("Identity") + ylab("") + ggtitle("STRA8") + theme(legend.position = "none")
```

```{r}
for(i in markers){
jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin_jitter/Mouse.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.mou, features.plot = i, ident.include = c(1:6), cols.use = cols, point.size.use = 0.05)
print(p)
dev.off()

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin_jitter/Human.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.hum, features.plot = i, ident.include = c(1:6), cols.use = cols, point.size.use = 0.05)
print(p)
dev.off()

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin_jitter/Monkey.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.mac, features.plot = i, ident.include = c(1:6), cols.use = cols, point.size.use = 0.05)
print(p)
dev.off()
}
```

```{r fig.width=3, fig.height=2.5}
vlnplot_gg <- function(object, gene, adjust=1){
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
df <- data.frame(expr = object@data[gene,], ident = object@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width", adjust=adjust) + ylab("") + xlab("") + ggtitle(gene)  + scale_fill_manual(values = cols)+ theme(legend.position = "none")
}
```

```{r}
markers <- c("TSPAN33")
#markers <- c("NANOS3")
#markers <- c("FMR1NB")
#markers <- c("PIWIL4")
for(i in markers){
  jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Human.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- vlnplot_gg(spg.cca.hum, i, adjust = 3)
print(p)
dev.off()

  jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Monkey.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- vlnplot_gg(spg.cca.mac, i, adjust = 3)
print(p)
dev.off()

  jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Mouse.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- vlnplot_gg(spg.cca.mou, i, adjust = 3)
print(p)
dev.off()
}
```


```{r}
markers <- c("FMR1NB", "VCX", "VCX2", "VCX3A", "VCX3B","GFRA1","ZBTB16","LIN28A","ID4","ITGA6","UCHL1","STRA8","MKI67","DMRT1","SYCP2","HORMAD1","MEIOB","LY6K","TEX101","PIWIL4","MORC1","TCF3","CDK17","FMR1","L1TD1")
markers <- c("FMR1NB")
for(i in markers){
jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Mouse.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.mou, features.plot = i, ident.include = c(1:6), cols.use = cols, point.size.use = 0, size.x.use = 0)
print(p)
dev.off()

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Human.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.hum, features.plot = i, ident.include = c(1:6), cols.use = cols, point.size.use = 0, size.x.use = 0)
print(p)
dev.off()

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Monkey.spg.cca","_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.mac, features.plot = i, ident.include = c(1:6), cols.use = cols, point.size.use = 0, size.x.use = 0)
print(p)
dev.off()
}
```


```{r}
hum.markers <- c("VCX", "VCX2", "VCX3A", "VCX3B")
mac.markers <- c("ENSMMUG00000047834","ENSMMUG00000039934","ENSMMUG00000043376")
```

Add hum.markers for human
```{r}
spg.cca.hum@meta.data$VCX <- 0
spg.cca.hum@meta.data[colnames(dge.hum.spg.full@data),]$VCX <- dge.hum.spg.full@data["VCX",]
spg.cca.hum@meta.data$VCX2 <- 0
spg.cca.hum@meta.data[colnames(dge.hum.spg.full@data),]$VCX2 <- dge.hum.spg.full@data["VCX2",]
spg.cca.hum@meta.data$VCX3A <- 0
spg.cca.hum@meta.data[colnames(dge.hum.spg.full@data),]$VCX3A <- dge.hum.spg.full@data["VCX3A",]
spg.cca.hum@meta.data$VCX3B <- 0
spg.cca.hum@meta.data[colnames(dge.hum.spg.full@data),]$VCX3B <- dge.hum.spg.full@data["VCX3B",]
```

```{r fig.width=3, fig.height=2.5}
vlnplot_gg <- function(object, gene, adjust=1){
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
df <- data.frame(expr = object@meta.data[,gene], ident = object@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width", adjust=adjust) + ylab("") + xlab("") + ggtitle(gene)  + scale_fill_manual(values = cols)+ theme(legend.position = "none")
}
```

```{r}
for(i in hum.markers){
jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Human.spg.cca_",i,"_violin.jpeg"),height=850,width=800,res=300)
p <- vlnplot_gg(spg.cca.hum, i, adjust = 1)
print(p)
dev.off()
}
```


Add KIT for humand and mouse
```{r}
spg.cca.hum@meta.data$KIT <- 0
spg.cca.hum@meta.data[colnames(dge.hum.spg.full@data),]$KIT <- dge.hum.spg.full@data["KIT",]
```

```{r}
spg.cca.mou@meta.data$Kit <- 0
spg.cca.mou@meta.data[colnames(dge.mou.spg.norm),]$Kit <- log(dge.mou.spg.norm["Kit",]+1)
```

```{r fig.width=3, fig.height=2.5}
vlnplot_feature <- function(object, feature){
cols <- c("#477900",
"#9378ff",
"#01ce74",
"#ff2b85",
"#b0d266",
"#733697",
"#ff9c1d",
"#0190c8",
"#ffa19e")
df <- data.frame(expr = object@meta.data[,feature], ident = object@meta.data$res.0.6.ord)
df <- df[df$ident %in% c(1:6),]
ggplot(df, aes(x=ident,y=expr, fill=ident))+geom_violin(scale = "width", adjust=1) + ylab("") + xlab("") + ggtitle(feature)  + scale_fill_manual(values = cols)+ theme(legend.position = "none")
}
```

```{r}
jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Mouse.spg.cca_Kit_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.mou, features.plot = "Kit", ident.include = c(1:6), cols.use = cols, point.size.use = 0, size.x.use = 0)
print(p)
dev.off()

jpeg(paste0("~/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/violin/Human.spg.cca_KIT_violin.jpeg"),height=850,width=800,res=300)
p <- VlnPlot(spg.cca.hum, features.plot = "KIT", ident.include = c(1:6), cols.use = cols, point.size.use = 0, size.x.use = 0)
print(p)
dev.off()
```


##. Find Markers only for the 6 clusters -- 20190907

```{r}
spg.cca.6 <- SubsetData(spg.cca,ident.use = c(1:6))
```


###. Conserved markers
```{r}
markers.6clus <- list()
for(i in c(1:6)){
  markers.6clus[[i]] <- FindConservedMarkers(spg.cca.6, ident.1 = i, grouping.var = "species", logfc.threshold = 0.25, min.pct = 0.1, only.pos = TRUE)
}
```

Write to excel
```{r}
for(i in c(1:6)){
if(i==1){
  write.xlsx2(x = markers.6clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_6Clusters_Markers.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = markers.6clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_6Clusters_Markers.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```

###findmarkers
```{r}
markers.sep.6clus <- list()
for(i in c(1:6)){
  markers.sep.6clus[[i]] <- FindMarkers(spg.cca.6, ident.1 = i, grouping.var = "species", logfc.threshold = 0.25, min.pct = 0.1, only.pos = TRUE)
}
```

```{r}
for(i in c(1:6)){
if(i==1){
  write.xlsx2(x = markers.sep.6clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_6Clusters_Markers_allmerge.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = markers.sep.6clus[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/SPG_markers/SPG_6Clusters_Markers_allmerge.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```
