==> germcells/14_Germ_Clustering_Monkey.Rmd <==
Germ cell independent clustering for monkey

```{r}
dge.mac.spg.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mac.spg.full.raw.rds")
dge.mac.germ.raw <- readRDS("~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mac.germ.raw.rds")
```

```{r}
dge.mac.raw <- cbind(dge.mac.spg.raw[,!colnames(dge.mac.spg.raw) %in% colnames(dge.mac.germ.raw)], dge.mac.germ.raw)
```


```{r}
dge.mac <- CreateSeuratObject(raw.data = dge.mac.raw)
dge.mac <- NormalizeData(object = dge.mac, normalization.method = "LogNormalize", 
    scale.factor = 10000)
```

```{r}
dge.mac <- FindVariableGenes(object = dge.mac,mean.function = ExpMean, dispersion.function = LogVMR, x.low.cutoff = 0.0125, y.cutoff = 1, x.high.cutoff = 3)
batchs <- as.vector(sapply(colnames(dge.mac.raw),function(x) strsplit(x,'_')[[1]][1]))
batchs[batchs %in% c("Monkey1", "Monkey2")] = "Monkey1"
batchs[batchs %in% c("Monkey3")] = "Monkey2"
batchs[batchs %in% c("Monkey4", "Monkey5")] = "Monkey3"
batchs[batchs %in% c("Monkey6", "Monkey7")] = "Monkey4"
batchs[batchs %in% c("Monkey8", "Monkey9")] = "Monkey5"
dge.mac@meta.data$indi = batchs
mac.spg <- dge.mac@cell.names %in% colnames(dge.mac.spg.raw)
dge.mac@meta.data$spg <- mac.spg
dge.mac <- ScaleData(object = dge.mac,vars.to.regress = c("nUMI"))
dge.mac <- RunPCA(object = dge.mac, pc.genes = dge.mac@var.genes)
dge.mac <- RunTSNE(object = dge.mac, dims.use = 1:15)
dge.mac <- RunUMAP(object = dge.mac, dims.use = 1:15)
```

```{r fig.width=6,fig.height=4.5}
PCAPlot(object = dge.mac, dim.1 = 1, dim.2 = 2)
PCAPlot(object = dge.mac, dim.1 = 1, dim.2 = 2, group.by="indi")
PCElbowPlot(object = dge.mac)
```

```{r fig.width=6,fig.height=4.5}
TSNEPlot(object = dge.mac, group.by="indi")
TSNEPlot(object = dge.mac)
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE)
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="indi")
```


```{r fig.width=6,fig.height=4.5}
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="indi",cells.highlight = dge.mac@cell.names[dge.mac@meta.data$indi=="Monkey1"])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="indi",cells.highlight = dge.mac@cell.names[dge.mac@meta.data$indi=="Monkey2"])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="indi",cells.highlight = dge.mac@cell.names[dge.mac@meta.data$indi=="Monkey3"])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="indi",cells.highlight = dge.mac@cell.names[dge.mac@meta.data$indi=="Monkey4"])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="indi",cells.highlight = dge.mac@cell.names[dge.mac@meta.data$indi=="Monkey5"])
```
```{r}
dge.mac <- FindClusters(object = dge.mac, reduction.type = "pca", dims.use = 1:15, save.SNN = TRUE, resolution = 0.2)
```

```{r fig.width=6, fig.height=4.5}
dge.mac@meta.data$res.0.2.ord <- plyr::mapvalues(dge.mac@meta.data$res.0.2, c(5,4,6,2,3,1,0),c(1,2,3,4,5,6,7))
TSNEPlot(object = dge.mac, do.return = TRUE, pt.size = 1, group.by = "res.0.2.ord", colors.use = cols[1:7])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="res.0.2.ord", cols.use = cols[1:7])
DimPlot(object = dge.mac, reduction.use = "pca", do.return = TRUE, group.by="res.0.2.ord", cols.use = cols[1:7])
```

```{r fig.width=6, fig.height=4}
df <- data.frame(dge.mac@dr$umap@cell.embeddings)
df <- cbind(df, dge.mac@meta.data$res.0.2.ord)
colnames(df) <- c("UMAP1","UMAP2","Cluster")
ggplot(data = df) + geom_point(aes(x=-UMAP2, y=-UMAP1, col=Cluster)) + scale_color_manual(values = cols[1:7]) +  guides(colour = guide_legend(override.aes = list(size=5))) + theme_bw() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        panel.background = element_rect(colour = "black"),
        legend.title=element_blank(),
        axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank())
#DimPlot(object = dge.mac, reduction.use = "umap",do.return = TRUE, group.by="res.0.2.ord", cols.use = cols[1:7])
```


```{r fig.width=5, fig.height=4}
dge.mac.7.cen <- matrix(0, nrow = dim(dge.mac@data)[1], ncol = 7)
for(i in c(1:7)){
  mac.i.cells <- dge.mac@cell.names[dge.mac@meta.data$res.0.2.ord==i]
  mac.i.pro <- apply(exp(dge.mac@data[,mac.i.cells])-1,1,mean)
  dge.mac.7.cen[,i] <- log(mac.i.pro+1)
}
dge.mac.7.cen.cor <- cor(dge.mac.7.cen, method = "spearman")
rownames(dge.mac.7.cen.cor) <- c(1:7)
colnames(dge.mac.7.cen.cor) <- c(1:7)

Heatmap(dge.mac.7.cen.cor, cluster_rows = FALSE, cluster_columns = FALSE)
```

```{r}
dge.mac <- FindClusters(object = dge.mac, reduction.type = "pca", dims.use = 1:15, save.SNN = TRUE, resolution = 0.3)
```

```{r fig.width=6, fig.height=4.5}
dge.mac@meta.data$res.0.3.ord <- plyr::mapvalues(dge.mac@meta.data$res.0.3, c(6,5,7,2,4,1,8,3,0),c(1,2,3,4,5,6,7,8,9))
TSNEPlot(object = dge.mac, do.return = TRUE, pt.size = 1, group.by = "res.0.3.ord", colors.use = cols[1:9])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="res.0.3.ord", cols.use = cols[1:9])
DimPlot(object = dge.mac, reduction.use = "pca", do.return = TRUE, group.by="res.0.3.ord", cols.use = cols[1:9])
```

```{r fig.width=5, fig.height=4}
dge.mac.9.cen <- matrix(0, nrow = dim(dge.mac@data)[1], ncol = 9)
for(i in c(1:9)){
  mac.i.cells <- dge.mac@cell.names[dge.mac@meta.data$res.0.3.ord==i]
  mac.i.pro <- apply(exp(dge.mac@data[,mac.i.cells])-1,1,mean)
  dge.mac.9.cen[,i] <- log(mac.i.pro+1)
}
dge.mac.9.cen.cor <- cor(dge.mac.9.cen, method = "spearman")
rownames(dge.mac.9.cen.cor) <- c(1:9)
colnames(dge.mac.9.cen.cor) <- c(1:9)

Heatmap(dge.mac.9.cen.cor, cluster_rows = FALSE, cluster_columns = FALSE)
```

```{r}
dge.mac <- FindClusters(object = dge.mac, reduction.type = "pca", dims.use = 1:5, save.SNN = TRUE, resolution = 0.5)
```

```{r fig.width=6, fig.height=4.5}
#dge.mac@meta.data$res.0.3.ord <- plyr::mapvalues(dge.mac@meta.data$res.0.3, c(6,5,7,2,4,1,8,3,0),c(1,2,3,4,5,6,7,8,9))
TSNEPlot(object = dge.mac, do.return = TRUE, pt.size = 1, group.by = "res.0.5", colors.use = cols[1:12])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="res.0.5", cols.use = cols[1:12])
DimPlot(object = dge.mac, reduction.use = "pca", do.return = TRUE, group.by="res.0.5", cols.use = cols[1:12])
```

```{r}
dge.mac <- FindClusters(object = dge.mac, reduction.type = "pca", dims.use = 1:5, save.SNN = TRUE, resolution = 0.3)
```

```{r fig.width=6, fig.height=4.5}
dge.mac@meta.data$res.0.3.ord <- plyr::mapvalues(dge.mac@meta.data$res.0.3, c(6,7,2,4,5,3,1,0),c(1,2,3,4,5,6,7,8))
TSNEPlot(object = dge.mac, do.return = TRUE, pt.size = 1, group.by = "res.0.3.ord", colors.use = cols[1:8])
DimPlot(object = dge.mac, reduction.use = "umap", do.return = TRUE, group.by="res.0.3.ord", cols.use = cols[1:8])
DimPlot(object = dge.mac, reduction.use = "pca", do.return = TRUE, group.by="res.0.3.ord", cols.use = cols[1:8])
```


```{r fig.width=5, fig.height=4}
dge.mac.8.cen <- matrix(0, nrow = dim(dge.mac@data)[1], ncol = 8)
for(i in c(1:8)){
  mac.i.cells <- dge.mac@cell.names[dge.mac@meta.data$res.0.3.ord==i]
  mac.i.pro <- apply(exp(dge.mac@data[,mac.i.cells])-1,1,mean)
  dge.mac.8.cen[,i] <- log(mac.i.pro+1)
}
dge.mac.8.cen.cor <- cor(dge.mac.8.cen, method = "spearman")
rownames(dge.mac.8.cen.cor) <- c(1:8)
colnames(dge.mac.8.cen.cor) <- c(1:8)

Heatmap(dge.mac.8.cen.cor, cluster_rows = FALSE, cluster_columns = FALSE)
```

```{r fig.width=10, fig.height=15}
par(mfrow=c(6,4))
for(i in markers[markers %in% rownames(dge.mac.raw)]){
  plotgene(dge.mac,i)
}
```

```{r fig.width=10, fig.height=6}
par(mfrow=c(2,3))
marker2 <- c("BEND2", "DMRT1","HORMAD1","PIWIL1", "ACRV1","PRM2")
for(i in marker2){
plotgene(dge.mac,i)
}
```


```{r fig.width=10, fig.height=15}
VlnPlot(dge.mac, features.plot = markers[markers %in% rownames(dge.mac.raw)], nCol = 4, group.by = "res.0.2.ord", point.size.use = 0)
```



```{r}
saveRDS(dge.mac, file = "~/storage/HumanMacaqueMouseCompAnaysis/interdata/dge.mac.germ.seurat.rds")
```

###. Call Markers

```{r}
dge.mac <- SetIdent(dge.mac, ident.use = dge.mac@meta.data$res.0.2.ord)
```

```{r}
mac.markers <- list()
for(i in c(1:7)){
  tmp <- FindMarkers(dge.mac, ident.1 = i, min.pct = 0.2, logfc.threshold = 0.7,only.pos = TRUE)
  tmp <- tmp[tmp$p_val_adj < 0.01, ]
  mac.markers[[i]] <- tmp
}
```

```{r}
library(xlsx)
```


```{r}
for(i in c(1:7)){
if(i==1){
  write.xlsx2(x = mac.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/Germ_Clusters/Mac_Germ_7Clusters_Markers.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = FALSE)
}else{
  write.xlsx2(x = mac.markers[[i]], file = "/home/xianingz/storage/HumanMacaqueMouseCompAnaysis/interdata/Germ_Clusters/Mac_Germ_7Clusters_Markers.xlsx", sheetName = paste0("Cluster", i), col.names = TRUE, row.names = TRUE, append = TRUE)
}
}
```
